+++ src/SFA.DAS.ProviderCommitments/SFA.DAS.ProviderCommitments.Web.UnitTests/Models/ApprenticeIndexViewModelTests.cs
@@ -1,11 +1,12 @@
 ﻿using NUnit.Framework;
 using SFA.DAS.ProviderCommitments.Web.Models;
+using SFA.DAS.ProviderCommitments.Web.Models.Apprentice;
 using SFA.DAS.Testing.AutoFixture;
 
 namespace SFA.DAS.ProviderCommitments.Web.UnitTests.Models
 {
     [TestFixture]
-    public class ManageApprenticesModelTests
+    public class ApprenticeIndexViewModelTests
     {
         [Test]
         [MoqInlineAutoData(true, "das-table__sort das-table__sort--desc")]
@@ -13,7 +14,7 @@ namespace SFA.DAS.ProviderCommitments.Web.UnitTests.Models
         public void ThenTheSortByHeaderClassNameIsSetCorrectly(
             bool isReverse,
             string expected,
-            ManageApprenticesViewModel model)
+            IndexViewModel model)
         {
             //Arrange
             model.SortedByHeaderClassName = "";
